Add file extension to bundles
A bundle have now a file extension ('js' for JavaScript bundles for
example). The  extension is used to build URLs to bundles and assets.
This is needed when you use Frozen-Flask (or similar tools) to build a
static version of your website. This is also the first step to
statically build bundles and assets in an application deployment
scenario.